frontend components tests

+++ packages/frontend/src/components/EventsTable/EventsTable.tsx
@@ -16,17 +16,19 @@ import { ErrorBox, Spinner, Icon } from "@ui";
 let showedAboutRandomizedUpdateAlert = false;
 
 const showAboutRandomizedUpdateAlert = () => {
-  if (!showedAboutRandomizedUpdateAlert) {
-    alert(
-      [
-        `NOTE:`,
-        `Updates could be handled inside table with inline inputs and debounce.`,
-        `I could also open a modal that allows editing or pass this data to part of the app where events are created.`,
-        `I dint't want to add to much stuff to this little demo app, so update will just randomize the event.`,
-      ].join("\n\n")
-    );
+  if (process.env.NODE_ENV !== "test") {
+    if (!showedAboutRandomizedUpdateAlert) {
+      alert(
+        [
+          `NOTE:`,
+          `Updates could be handled inside table with inline inputs and debounce.`,
+          `I could also open a modal that allows editing or pass this data to part of the app where events are created.`,
+          `I dint't want to add to much stuff to this little demo app, so update will just randomize the event.`,
+        ].join("\n\n")
+      );
 
-    showedAboutRandomizedUpdateAlert = true;
+      showedAboutRandomizedUpdateAlert = true;
+    }
   }
 };
 
@@ -41,8 +43,8 @@ const makeRandomString = (length: number): string =>
 /* -------------------------------------------------------------------------- */
 
 interface IEventsTableRowProps extends IEventDto {
-  onDeleteEvent: (id: IEventId) => void;
-  onUpdateEvent: (id: IEventId, changes: IUpdateEventDto) => void;
+  onDeleteEvent?: (id: IEventId) => void;
+  onUpdateEvent?: (id: IEventId, changes: IUpdateEventDto) => void;
 }
 
 /*
@@ -51,43 +53,47 @@ interface IEventsTableRowProps extends IEventDto {
   which rows have changed, and it won't have to render every row again if some other
   row is updated, deleted or added.
 
-  This could be in a separate file if it had some more logic.
+  I would save it in a separate file if it had more logic.
 */
 export const EventsTableRow: React.VFC<IEventsTableRowProps> = React.memo((props) => {
   const { id, date, email, firstName, lastName, onDeleteEvent, onUpdateEvent } = props;
 
   /*
     NOTE:
-    Buttons are not some complex components so I don't see any reasons
+    <button>s are not some complex components so I don't see any reasons
     to use useCallback here. I left it like this since it's more readable.
   */
   const handleDelete = () => {
-    onDeleteEvent(id);
+    if (onDeleteEvent) {
+      onDeleteEvent(id);
+    }
   };
 
   const handleUpdate = () => {
-    showAboutRandomizedUpdateAlert();
+    if (onUpdateEvent) {
+      showAboutRandomizedUpdateAlert();
 
-    onUpdateEvent(id, {
-      firstName: makeRandomString(8),
-      lastName: makeRandomString(8),
-      email: `${makeRandomString(8)}@${makeRandomString(8)}.com`,
-      date: new Date(Math.random() * 10000000000000).toISOString(),
-    });
+      onUpdateEvent(id, {
+        firstName: makeRandomString(8),
+        lastName: makeRandomString(8),
+        email: `${makeRandomString(8)}@${makeRandomString(8)}.com`,
+        date: new Date(Math.random() * 10000000000000).toISOString(),
+      });
+    }
   };
 
   return (
-    <tr>
+    <tr data-eventid={id}>
       <td>{id}</td>
       <td>{firstName}</td>
       <td>{lastName}</td>
       <td>{email}</td>
       <td>{date}</td>
       <td className={styles.actionButtonsCell}>
-        <button className={styles.actionButton} onClick={handleDelete}>
+        <button aria-label="delete" className={styles.actionButton} onClick={handleDelete}>
           <Icon.Trash />
         </button>
-        <button className={styles.actionButton} onClick={handleUpdate}>
+        <button aria-label="update" className={styles.actionButton} onClick={handleUpdate}>
           <Icon.Pencil />
         </button>
       </td>
@@ -99,24 +105,31 @@ export const EventsTableRow: React.VFC<IEventsTableRowProps> = React.memo((props
 /*                               MAIN COMPONENT                               */
 /* -------------------------------------------------------------------------- */
 
-interface IEventsTableProps {
+/* 
+  NOTE:
+  Messages exported for tests.
+  If application was localized, test would probably use data
+  from a json or some external provider in both places.
+*/
+export const ERROR_TITLE = "Failed to fetch events data";
+export interface IEventsTableProps {
   className?: string;
-  isFetching: boolean;
+  isFetching?: boolean;
   events: IEventsArrayDto;
-  error: string | null;
-  onDeleteEvent: (id: IEventId) => void;
-  onUpdateEvent: (id: IEventId, changes: IUpdateEventDto) => void;
+  error?: string | null;
+  onDeleteEvent?: (id: IEventId) => void;
+  onUpdateEvent?: (id: IEventId, changes: IUpdateEventDto) => void;
 }
 
 export const EventsTable: React.VFC<IEventsTableProps> = React.memo((props) => {
-  const { className, isFetching, events, error, onDeleteEvent, onUpdateEvent } = props;
+  const { className, isFetching = false, events, error, onDeleteEvent, onUpdateEvent } = props;
 
   let overlayContentJsx: React.ReactNode | undefined;
 
   if (isFetching && events.length === 0) {
     overlayContentJsx = <Spinner size={60} />;
   } else if (error) {
-    overlayContentJsx = <ErrorBox title="Failed to fetch events data" message={error} />;
+    overlayContentJsx = <ErrorBox title={ERROR_TITLE} message={error} />;
   }
 
   return (
